Refactor

+++ src/models.py
@@ -2,10 +2,10 @@
 
 import time
 import numpy as np
-from lib.utils import *
+from utils import *
 from collections import Counter
-from lib.multiprototype import *
-from lib.feature_data import *
+from multiprototype import *
+from feature_data import *
 import torch.nn as nn
 import torch.nn.functional as F
 from torch import optim
@@ -52,7 +52,7 @@ def add_models_args(parser):
     parser.add_argument('--mu2', type=float, default=None, help='mu_cont (Talukdar and Kramer 2009)')
     parser.add_argument('--mu3', type=float, default=None, help='mu_abdn (Talukdar and Kramer 2009)')
     parser.add_argument('--mu4', type=float, default=None, help='ModAds NNk (Rosenfeld and Erk 2019)')
-    parser.add_argument('--nnk', type=float, default=None, help='ModAds equal/decay n (Rosenfeld and Erk 2019')
+    parser.add_argument('--nnk', type=int, default=None, help='ModAds equal/decay n (Rosenfeld and Erk 2019')
 
 
 # class DumbClassifier(object):
